refactor: tests import the lib, not the CLI entrypoint
Three of the four tests were reaching through an entrypoint for helpers that
already lived in a lib. timer's formatCountdown/generateId and time-machine's
runTimeMachine did not, so they moved to src/<tool>/lib/.
