+++ src/dev-dashboard/index.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, test } from "bun:test";
 
 describe("killChild", () => {
     test("escalates to SIGKILL if the child ignores SIGTERM within the grace period", async () => {
-        const { killChild } = await import("./index");
+        const { killWithEscalation: killChild } = await import("@genesiscz/utils/process/killWithEscalation");
 
         const proc = Bun.spawn(["sh", "-c", "trap '' TERM; sleep 30"]);
         await new Promise((r) => setTimeout(r, 100));
