+++ src/timer/index.ts
@@ -11,6 +11,7 @@ import { Storage } from "@genesiscz/utils/storage/storage";
 import { formatTable } from "@genesiscz/utils/table";
 import { Command } from "commander";
 import pc from "picocolors";
+import { formatCountdown, generateId } from "./lib/format";
 
 // ============================================
 // Types
@@ -89,30 +90,9 @@ async function removeActiveTimer(id: string): Promise<void> {
     });
 }
 
-export function generateId(): string {
-    return Date.now().toString(36) + Math.random().toString(36).slice(2, 6);
-}
-
 // ============================================
 // Countdown display
 // ============================================
-
-export function formatCountdown(remainingMs: number): string {
-    if (remainingMs <= 0) {
-        return "00:00";
-    }
-
-    const totalSeconds = Math.ceil(remainingMs / 1000);
-    const hours = Math.floor(totalSeconds / 3600);
-    const minutes = Math.floor((totalSeconds % 3600) / 60);
-    const seconds = totalSeconds % 60;
-
-    if (hours > 0) {
-        return `${String(hours).padStart(2, "0")}:${String(minutes).padStart(2, "0")}:${String(seconds).padStart(2, "0")}`;
-    }
-
-    return `${String(minutes).padStart(2, "0")}:${String(seconds).padStart(2, "0")}`;
-}
 
 async function runForegroundTimer(opts: TimerOptions): Promise<void> {
     const { durationMs, title, notify, say, repeat } = opts;
